fix(mount): --ro-bind 非 root 下 remount EPERM + 挂载 CLI 顺序
- ro-bind remount 现在带上源挂载的锁定 flags（source_mount_flags 从
  /proc/self/mountinfo 读 opts 映射为 MS_*），修复 userns 里移除
  nosuid/nodev 等锁定 flag 导致的 EPERM
- main.rs 用 indices_of 按 CLI 顺序合并 --bind/--ro-bind/--tmpfs，
  修复叠加可写子目录被后挂基座埋住（只读源码 + 可写构建产物模式）
- 测试：M3b ro-bind 非 root 回归、M3c ro-bind 基座 + 可写叠加回归
- docs: learned.md 与 mount-namespace design.md §8.5 更新为真实修复

+++ src/config.rs
@@ -536,7 +536,7 @@ mod tests {
         assert_eq!(spec.source, Some("/host/src".into()));
         assert_eq!(spec.target, "/container/dst");
         assert_eq!(spec.fstype, "none");
-        assert_eq!(spec.flags, (MS_BIND | MS_REC) as u64);
+        assert_eq!(spec.flags, MS_BIND | MS_REC);
         assert!(!spec.readonly);
     }
 
@@ -554,6 +554,6 @@ mod tests {
         assert!(spec.source.is_none());
         assert_eq!(spec.target, "/mnt/tmp");
         assert_eq!(spec.fstype, "tmpfs");
-        assert_eq!(spec.flags, (MS_NOSUID | MS_NODEV) as u64);
+        assert_eq!(spec.flags, MS_NOSUID | MS_NODEV);
     }
 }
